pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "cython<3.0",  # To solve and error in av\logging.pyx: https://github.com/SYSTRAN/faster-whisper/issues/560 (av is needed for whisper that's needed for nexaai)
]
build-backend = "setuptools.build_meta"

[project]
name = "ai-amy"
version = "0.0.1"
description = "aiAmy"
requires-python = ">=3.8"

dependencies = [
    "loguru==0.7.3",
    "Pillow==11.1.0",
    "Pydantic==2.11.1",
    "llama-cpp-python==0.3.8",
    "PyInstaller==6.12.0"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4"    
]

[project.scripts]
cli-tool = "ai_amy.main:run"

[tool.pip-tools]
extra-index-url = [
    "https://pypi.org/simple",
    "https://abetlen.github.io/llama-cpp-python/whl/cpu"
]